wip: add voice plugin
Add `byteorder`, `cpal`, and `futures` as dependencies.
Add `voice` feature to `Cargo.toml`.
Add `voice` module with handlers for the commands: `join`, `listen`, and
`listen`. These commands allow the bot to join a specified channel, and either
begin listening or stop listening to voice chat. It then attempts to perform TTS
on the captured voice data.